+++ tensor/preprocess.h
@@ -20,7 +20,7 @@ namespace ctce {
     /**
      * Generate anti-symmetry iterator
      */
-    void genAntiIter(const std::vector<Integer> &vtab, IterGroup<antisymm>& ext_itr, const Tensor& tC, const Tensor& tA, const Tensor& tB);
+    void genAntiIter(const std::vector<size_t> &vtab, IterGroup<antisymm>& ext_itr, const Tensor& tC, const Tensor& tA, const Tensor& tB);
 
 
   } // extern C
